EES-706 Fix FilterGroup grouping using equality comparer.
