+++ src/GovUk.Education.ExploreEducationStatistics.Data.Model/Filter.cs
@@ -1,6 +1,5 @@
 using System;
 using System.Collections.Generic;
-using System.ComponentModel.DataAnnotations.Schema;
 
 namespace GovUk.Education.ExploreEducationStatistics.Data.Model
 {
@@ -15,6 +14,8 @@ namespace GovUk.Education.ExploreEducationStatistics.Data.Model
         public ICollection<FilterGroup> FilterGroups { get; set; }
         public ICollection<FilterFootnote> Footnotes { get; set; }
 
+        public static IEqualityComparer<Filter> IdComparer { get; } = new IdEqualityComparer();
+
         public Filter()
         {
         }
@@ -27,6 +28,23 @@ namespace GovUk.Education.ExploreEducationStatistics.Data.Model
             Name = name;
             Subject = subject;
             FilterGroups = new List<FilterGroup>();
+        }
+
+        private sealed class IdEqualityComparer : IEqualityComparer<Filter>
+        {
+            public bool Equals(Filter x, Filter y)
+            {
+                if (ReferenceEquals(x, y)) return true;
+                if (ReferenceEquals(x, null)) return false;
+                if (ReferenceEquals(y, null)) return false;
+                if (x.GetType() != y.GetType()) return false;
+                return x.Id.Equals(y.Id);
+            }
+
+            public int GetHashCode(Filter obj)
+            {
+                return obj.Id.GetHashCode();
+            }
         }
     }
 }
