+++ src/GovUk.Education.ExploreEducationStatistics.Data.Model/FilterGroup.cs
@@ -1,6 +1,5 @@
 using System;
 using System.Collections.Generic;
-using System.ComponentModel.DataAnnotations.Schema;
 
 namespace GovUk.Education.ExploreEducationStatistics.Data.Model
 {
@@ -13,6 +12,8 @@ namespace GovUk.Education.ExploreEducationStatistics.Data.Model
         public ICollection<FilterItem> FilterItems { get; set; }
         public ICollection<FilterGroupFootnote> Footnotes { get; set; }
 
+        public static IEqualityComparer<FilterGroup> IdComparer { get; } = new IdEqualityComparer();
+
         public FilterGroup()
         {
         }
@@ -23,6 +24,23 @@ namespace GovUk.Education.ExploreEducationStatistics.Data.Model
             Filter = filter;
             Label = label;
             FilterItems = new List<FilterItem>();
+        }
+
+        private sealed class IdEqualityComparer : IEqualityComparer<FilterGroup>
+        {
+            public bool Equals(FilterGroup x, FilterGroup y)
+            {
+                if (ReferenceEquals(x, y)) return true;
+                if (ReferenceEquals(x, null)) return false;
+                if (ReferenceEquals(y, null)) return false;
+                if (x.GetType() != y.GetType()) return false;
+                return x.Id.Equals(y.Id);
+            }
+
+            public int GetHashCode(FilterGroup obj)
+            {
+                return obj.Id.GetHashCode();
+            }
         }
     }
 }
